+++ src/test/java/com/programmers/voucher/io/ConsoleMenuTest.java
@@ -1,9 +1,8 @@
-package com.programmers.voucher.controller;
+package com.programmers.voucher.io;
 
+import com.programmers.voucher.controller.VoucherConsoleController;
 import com.programmers.voucher.domain.Voucher;
 import com.programmers.voucher.enumtype.ConsoleCommandType;
-import com.programmers.voucher.io.Console;
-import com.programmers.voucher.service.VoucherService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -17,17 +16,18 @@ import java.util.UUID;
 
 import static com.programmers.voucher.testutil.VoucherTestUtil.createFixedVoucher;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-class ConsoleClientTest {
+class ConsoleMenuTest {
 
     @InjectMocks
-    private ConsoleClient consoleClient;
+    private ConsoleMenu consoleMenu;
 
     @Mock
-    private VoucherService voucherService;
+    private VoucherConsoleController voucherConsoleController;
 
     @Mock
     private Console console;
@@ -39,10 +39,10 @@ class ConsoleClientTest {
         given(console.input(anyString())).willReturn("fixed");
         given(console.intInput(anyString())).willReturn(10);
 
-        given(voucherService.createVoucher(any())).willReturn(UUID.randomUUID());
+        given(voucherConsoleController.createVoucher(any())).willReturn(UUID.randomUUID());
 
         //when
-        consoleClient.runClient();
+        consoleMenu.runClient();
 
         //then
     }
@@ -56,7 +56,7 @@ class ConsoleClientTest {
         //when
 
         //then
-        assertThatThrownBy(() -> consoleClient.runClient())
+        assertThatThrownBy(() -> consoleMenu.runClient())
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -70,7 +70,7 @@ class ConsoleClientTest {
         //when
 
         //then
-        assertThatThrownBy(() -> consoleClient.runClient())
+        assertThatThrownBy(() -> consoleMenu.runClient())
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -87,7 +87,7 @@ class ConsoleClientTest {
         //when
 
         //then
-        assertThatThrownBy(() -> consoleClient.runClient())
+        assertThatThrownBy(() -> consoleMenu.runClient())
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -99,10 +99,10 @@ class ConsoleClientTest {
         List<Voucher> testVouchers = List.of(fixedVoucher1, fixedVoucher2);
 
         given(console.inputInitialCommand()).willReturn(ConsoleCommandType.LIST);
-        given(voucherService.findVouchers()).willReturn(testVouchers);
+        given(voucherConsoleController.findVouchers()).willReturn(testVouchers);
 
         //when
-        consoleClient.runClient();
+        consoleMenu.runClient();
 
         //then
